LPS-106601 Unnecessary unboxing.

+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/background/task/CTServicePublisher.java
@@ -428,7 +428,7 @@ public class CTServicePublisher<T extends CTModel<T>> {
 		sb.append(" in (");
 
 		for (Serializable serializable : ctEntries.keySet()) {
-			sb.append((long)serializable);
+			sb.append(serializable);
 			sb.append(", ");
 		}
